Refactoring delete and ending the services tests

+++ src/main/java/br/com/gustavoakira/master/service/AddressService.java
@@ -38,7 +38,9 @@ public class AddressService {
 	}
 	
 	public String deleteById(Long id) {
-		addressRepository.deleteById(id);
+		if(getOne(id)!= null) {
+			addressRepository.deleteById(id);
+		}
 		return "ok";
 	}
 }
